Moved download function into FileInfo as a member
